ppc64: Quickfix for pointer2cell() and cell2pointer()
Add a uintptr_t cast as an interim solution to resolve errors for ppc64,
where pointer size is larger than cell size.

v2:
* Add comment on potential better fix.
  Postpone potential conversion to inline function until investigated.



git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@931 f158a5a8-5612-0410-a976-696ce0be7e32

+++ include/kernel/stack.h
@@ -31,8 +31,9 @@ typedef ucell phandle_t;
 
 
 #ifdef NATIVE_BITWIDTH_EQUALS_HOST_BITWIDTH
-#define pointer2cell(x) ((ucell)(x))
-#define cell2pointer(x) ((u8 *)(x))
+// XXX Check whether we can take the larger-than code path for ppc64 instead.
+#define pointer2cell(x) ((ucell)(uintptr_t)(x))
+#define cell2pointer(x) ((u8 *)(uintptr_t)(x))
 #endif
 #ifdef NATIVE_BITWIDTH_SMALLER_THAN_HOST_BITWIDTH
 #define pointer2cell(x) ((ucell)(((unsigned long)(x))-base_address))
